+++ src/soc/amd/common/block/include/amdblocks/root_complex.h
@@ -20,7 +20,7 @@ struct non_pci_mmio_reg {
 	unsigned long res_idx; /* Use NON_PCI_RES_IDX_AUTO or a specific resource index */
 };
 
-void read_non_pci_resources(struct device *domain, unsigned int *idx);
+void read_non_pci_resources(struct device *domain, unsigned long *idx);
 
 uint32_t get_iohc_misc_smn_base(struct device *domain);
 const struct non_pci_mmio_reg *get_iohc_non_pci_mmio_regs(size_t *count);
